Move passport strategy to util/user
